Fixing 32bit MSVC compilation
- flyby: fixed warnings about integral conversions

+++ hpx/performance_counters/registry.hpp
@@ -128,7 +128,7 @@ namespace hpx { namespace performance_counters
         ///        (milliseconds).
         counter_status create_statistics_counter(counter_info const& info,
             std::string const& base_counter_name,
-            std::vector<std::int64_t> const& parameters,
+            std::vector<std::size_t> const& parameters,
             naming::gid_type& id, error_code& ec = throws);
 
         /// \brief Create a new arithmetics performance counter instance based
